pyproject.toml
[tool.poetry]
name = "brewblox_usb_proxy"
version = "0.1.0"
description = "USB communication proxy for Brewblox devices"
authors = ["BrewPi B.V. <development@brewpi.com>"]
license = "GPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<4"
fastapi = "^0.109.2"
uvicorn = { extras = ["standard"], version = "^0.27.1" }
pydantic-settings = "^2.1.0"
pyserial = "^3.5"

[tool.poetry.group.dev.dependencies]
pytest-mock = "*"
flake8 = "*"
autopep8 = "*"
flake8-quotes = "*"
pytest = "*"
invoke = "*"
flake8-pyproject = "*"
pytest-asyncio = "*"
pytest-httpx = "*"
asgi-lifespan = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = """
    --ignore=app/
    --durations=3
    """

[tool.flake8]
max-line-length = 120
exclude = ".venv"

[tool.autopep8]
max-line-length = 120
